requirements.txt
boto3==1.26.62
botocore==1.29.62
jmespath==1.0.1
polars==0.16.1
python-dateutil==2.8.2
python-dotenv==0.21.1
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
